+++ packages/ionbase-ui/src/components/index.ts
@@ -1,5 +1,11 @@
 export { Button } from './Button.js';
 export type { ButtonProps } from './Button.js';
+export { Pagination } from './Pagination.js';
+export type {
+  PaginationProps,
+  PaginationType,
+  PaginationSize,
+} from './Pagination.js';
 export { Tabs, TabItem } from './Tabs.js';
 export type { TabsProps, TabsType, TabsSize } from './Tabs.js';
 export { Alert } from './Alert.js';